requirements.txt
matplotlib
numpy
scipy
pynrrd
scikit-image==0.16.2
xmltodict
h5py
pandas
sphinx==2.4.4
rst2pdf
m2r
mistune==0.8.4
coverage
deprecated
torch